select opportunity for courier delivery
