--- src/containers/checkout/checkout-form/delivery/courier/index.js
@@ -1,3 +1,0 @@
-import Courier from './courier';
-
-export default Courier;
