+++ src/containers/checkout/checkout-form/delivery/tests/delivery.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import Delivery from '../delivery';
-import Courier from '../courier/courier';
+import UkrpostAndCourier from '../ukrpost-and-courier';
 import NovaPost from '../nova-post';
 
 let wrapper;
@@ -9,8 +9,8 @@ beforeEach(() => {
 });
 
 describe('Delivery component tests', () => {
-  it('should render one <Courier>', () => {
-    expect(wrapper.find(Courier)).toHaveLength(1);
+  it('should render one <UkrpostAndCourier>', () => {
+    expect(wrapper.find(UkrpostAndCourier)).toHaveLength(1);
   });
   it('should render zero <NovaPost>', () => {
     expect(wrapper.find(NovaPost)).toHaveLength(0);
