+++ src/containers/checkout/checkout-form/delivery/ukrpost-and-courier/index.js
@@ -1,0 +1,3 @@
+import UkrpostAndCourier from './ukrpost-and-courier';
+
+export default UkrpostAndCourier;
