+++ src/containers/checkout/checkout-form/delivery/ukrpost-and-courier/tests/ukrpost-and-courier.spec.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import Autocomplete from '@material-ui/lab/Autocomplete';
 import { useQuery } from '@apollo/client';
-import UkrPost from '../ukrpost';
+import UkrpostAndCourier from '../ukrpost-and-courier';
 
-jest.mock('../ukrpost.styles.js', () => ({
+jest.mock('../ukrpost-and-courier.styles.js', () => ({
   useStyles: () => ({})
 }));
 jest.mock('react-redux');
@@ -19,22 +19,26 @@ let wrapper;
 
 const props = {
   isLightTheme: true,
+  language: 1,
   values: {
     firstName: '',
     lastName: '',
     email: '',
     phoneNumber: null
   },
+  handleChange: jest.fn(),
   touched: {
     region: 'Житомирська',
     district: 'Житомирський',
     city: 'Житомир',
+    street: 'Шевченка',
     courierOffice: '1'
   },
   errors: {
     region: 'Cannot be empty',
     district: 'Cannot be empty',
     city: 'Cannot be empty',
+    street: 'Cannot be empty',
     courierOffice: 'Cannot be empty'
   },
   setFieldValue: jest.fn()
@@ -43,14 +47,14 @@ useQuery.mockImplementation(() => ({
   ...useQueryData
 }));
 
-describe('UkrPost component tests', () => {
-  wrapper = mount(<UkrPost {...props} />);
+describe('UkrpostAndCourier component tests', () => {
+  wrapper = mount(<UkrpostAndCourier {...props} />);
 
-  it('Should render UkrPost', () => {
+  it('Should render UkrpostAndCourier', () => {
     expect(wrapper).toBeDefined();
   });
 
-  it('UkrPost inputs test', () => {
+  it('UkrpostAndCourier inputs test', () => {
     wrapper.find(Autocomplete).forEach((i) => {
       i.props().onChange('event', 'test');
       i.props().onChange('event');
@@ -61,17 +65,18 @@ describe('UkrPost component tests', () => {
       expect(i.props().options).toEqual([]);
     });
   });
-  it('should add data to UkrPost fields', () => {
+  it('should add data to UkrpostAndCourier fields', () => {
     useQuery.mockImplementation(() => ({
       ...useQueryData,
       data: {
         getUkrPoshtaRegions: ['test', 'test2'],
         getUkrPoshtaDistrictsByRegionId: ['test', 'test2'],
         getUkrPoshtaCitiesByDistrictId: ['test', 'test2'],
+        getUkrPoshtaStreetsByCityId: ['test', 'test2'],
         getUkrPoshtaPostofficesCityId: ['test', 'test2']
       }
     }));
-    wrapper = mount(<UkrPost {...props} />);
+    wrapper = mount(<UkrpostAndCourier {...props} />);
     wrapper.find(Autocomplete).forEach((i) => {
       expect(i.props().options).toEqual(['test', 'test2']);
     });
