+++ src/containers/checkout/checkout-form/delivery/ukrpost-and-courier/ukrpost-and-courier.styles.js
@@ -1,7 +1,7 @@
 import { makeStyles } from '@material-ui/core/styles';
 
 export const useStyles = makeStyles(({ palette }) => ({
-  courierContainer: {
+  addressContainer: {
     display: 'flex',
     flexDirection: 'column',
     justifyContent: 'space-around',
@@ -11,10 +11,6 @@ export const useStyles = makeStyles(({ palette }) => ({
   inputWrapper: {
     width: '100%',
     marginLeft: '10%'
-  },
-  error: {
-    color: '#e60000',
-    marginBottom: '2%'
   },
   textField: {
     width: 600,
@@ -34,13 +30,35 @@ export const useStyles = makeStyles(({ palette }) => ({
       width: '80%'
     }
   },
-  courierTitle: {
+  addressTitle: {
     fontWeight: 700,
     fontSize: 23,
     color: palette.textColor,
     marginLeft: '10%',
     '@media (max-width: 920px)': {
       width: '100%'
+    }
+  },
+  error: {
+    color: '#e60000',
+    marginBottom: '2%'
+  },
+  dataInput: {
+    width: 600,
+    marginBottom: '1%',
+    '& label.Mui-focused': {
+      color: palette.textColor
+    },
+    '& .MuiOutlinedInput-root': {
+      '& fieldset': {
+        borderColor: palette.textColor
+      },
+      '&.Mui-focused fieldset': {
+        borderColor: palette.textColor
+      }
+    },
+    '@media (max-width: 768px)': {
+      width: '80%'
     }
   }
 }));
