--- src/containers/checkout/checkout-form/delivery/ukrpost/index.js
@@ -1,3 +1,0 @@
-import UkrPost from './ukrpost';
-
-export default UkrPost;
