+++ src/containers/checkout/delivery-type/delivery-type.styles.js
@@ -1,6 +1,17 @@
 import { makeStyles } from '@material-ui/core/styles';
 
 export const useStyles = makeStyles(({ palette }) => ({
+  deliveryTypeContainer: {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'space-around',
+    alignItems: 'flex-start',
+    width: '100%'
+  },
+  inputWrapper: {
+    width: '100%',
+    marginLeft: '10%'
+  },
   btnWrapper: {
     display: 'flex',
     flexDirection: 'column',
@@ -44,6 +55,38 @@ export const useStyles = makeStyles(({ palette }) => ({
     marginLeft: '10%',
     marginBottom: 10,
     '@media (max-width: 920px)': {
+      width: '100%'
+    }
+  },
+  error: {
+    color: '#e60000',
+    marginBottom: '2%'
+  },
+  formControl: {
+    width: 600,
+    marginBottom: 10,
+    '& label.Mui-focused': {
+      color: palette.textColor
+    },
+    '& .MuiOutlinedInput-root': {
+      '& fieldset': {
+        borderColor: palette.textColor
+      },
+      '&.Mui-focused fieldset': {
+        borderColor: palette.textColor
+      }
+    },
+    '@media (max-width: 768px)': {
+      flexDirection: 'column',
+      width: '80%',
+      margin: 0
+    }
+  },
+  courierSelect: {
+    height: 50,
+    width: '100%',
+    marginBottom: 10,
+    '@media (max-width: 768px)': {
       width: '100%'
     }
   }
