+++ src/locales/ua/delivery.json
@@ -9,6 +9,8 @@
   "sunday": "Нд:  ",
   "noCity": "Жодного міста не знайдено",
   "noDepartment": "Жодного відділення не знайдено",
+  "street": "Вулиця",
+  "noStreet": "Жодної вулиці не знайдено",
   "deliveryAddress": "Адреса доставки",
   "city": "Місто",
   "department": "Відділення",
